backend/requirements.txt
# FastAPI & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Machine Learning
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
scipy>=1.11.0

# Data Handling
joblib>=1.3.0

# Development & Testing
pytest>=7.4.0
python-dotenv>=1.0.0
